slap a bunch of dry runs

+++ bakery.py
@@ -687,16 +687,28 @@ def enable_locales(to_en: list) -> None:
     if len(to_add):
         for i in to_add:  # Why fumble with stacked \\n? Just spam a bit.
             lp("Enabling:" + i)
-            subprocess.run(["sudo", "bash", "-c", "echo " + i + ">> /etc/locale.gen"])
+            cmd = ["sudo", "bash", "-c", "echo " + i + ">> /etc/locale.gen"]
+            if dryrun:
+                lp("Would have run: " + str(cmd))
+            else:
+                subprocess.run(cmd)
         lp("Generating locales")
-        subprocess.run(["sudo", "locale-gen"])
+        cmd = ["sudo", "locale-gen"]
+        if dryrun:
+            lp("Would have run: " + str(cmd))
+        else:
+            subprocess.run(cmd)
 
 
 def set_locale(locale: str) -> None:
     if locale not in locales(True):
         raise OSError("Locale not enabled!")
     lp("Setting locale to: " + locale)
-    subprocess.run(["sudo", "localectl", "set-locale", "LANG=" + locale])
+    cmd = ["sudo", "localectl", "set-locale", "LANG=" + locale]
+    if dryrun:
+        lp("Would have run: " + str(cmd))
+    else:
+        subprocess.run(cmd)
 
 
 def set_kb(locale: str) -> None:
@@ -759,12 +771,19 @@ def package_desc(packages: list) -> dict:
 
 def enable_services(services: list) -> None:
     try:
-        subprocess.run(["sudo", "systemctl", "enable", i])
+        cmd = ["sudo", "systemctl", "enable", i]
+        if dryrun:
+            lp("Would have run: " + str(cmd))
+        else:
+            subprocess.run(cmd)
     except:
         pass
 
 
 def setup_base() -> None:
+    if dryrun:
+        lp("Setup base skipped in dryrun")
+        return
     os.remove("/etc/sudoers.d/g_wheel")
     os.remove("/etc/polkit-1/rules.d/49-nopasswd_global.rules")
 
@@ -895,25 +914,37 @@ def adduser(username: str, passwd: str, uid, gid, shell: str, groups: list) -> N
     if gidc(gid):
         raise OSError("Used GID")
     lp("Making group " + username + " on gid " + gid)
-    subprocess.run(
-        ["sudo", "groupadd", username, "-g", gid]
-    )  # May silently fail, which is fine.
+    cmd = ["sudo", "groupadd", username, "-g", gid]
+    if dryrun:
+        lp("Would have run: " + str(cmd))
+    else:
+        subprocess.run(cmd)  # May silently fail, which is fine.
     lp("Adding user " + username + "on " + uid + ":" + gid + " with shell " + shell)
-    subprocess.run(
-        ["sudo", "useradd", "-N", username, "-u", uid, "-g", gid, "-m", "-s", shell]
-    )
+    cmd = ["sudo", "useradd", "-N", username, "-u", uid, "-g", gid, "-m", "-s", shell]
+    if dryrun:
+        lp("Would have run: " + str(cmd))
+    else:
+        subprocess.run(cmd)
     for i in groups:
         groupadd(username, i)
 
 
 def groupadd(username: str, group: str) -> None:
     lp("Adding " + username + " to group " + group)
-    subprocess.run(["sudo", "usermod", "-aG", username, group])
+    cmd = ["sudo", "usermod", "-aG", username, group]
+    if dryrun:
+        lp("Would have run: " + str(cmd))
+    else:
+        subprocess.run(cmd)
 
 
 def passwd(username: str, passwd: str) -> None:
     lp("Setting user " + username + " password")
-    subprocess.run(["sudo", "passwd", username], input=f"{passwd}\n{passwd}", text=True)
+    cmd = ["sudo", "passwd", username]
+    if dryrun:
+        lp("Would have run: " + str(cmd))
+    else:
+        subprocess.run(cmd, input=f"{passwd}\n{passwd}", text=True)
 
 
 # Main functions
